+++ index.js
@@ -10,6 +10,19 @@ module.exports = {
       files: ['*.less', '**/*.less'],
       customSyntax: 'postcss-less',
     },
+    {
+      files: [
+        '*.js',
+        '**/*.js',
+        '*.jsx',
+        '**/*.jsx',
+        '*.ts',
+        '**/*.ts',
+        '*.tsx',
+        '**/*.tsx',
+      ],
+      customSyntax: '@stylelint/postcss-css-in-js',
+    },
   ],
   ignoreFiles: ['build', 'coverage', 'dist', 'lib', 'node_modules'],
 };